user account CRUD system complete

+++ api/resetPasswordSystem.js
@@ -31,6 +31,8 @@ export default async function handler(request, response) {
             const token = receivedData.token;
             // Find the user's email associated with the provided token
             const passwordResetUserRecord = await passwordResetCollection.findOne({ token });
+            // Assign token in database to variable
+            const tokenInDatabase = passwordResetUserRecord.token;
 
             // If there is no email associated with the token, return an error
             if (!passwordResetUserRecord) {
@@ -38,17 +40,25 @@ export default async function handler(request, response) {
                 return;
             }
 
-            // Retrieve the user's email address if token matches
-            const email = passwordResetUserRecord.email;
-            // Hash the password entered into reset password
-            const hashedPassword = await bcrypt.hash(receivedData.password, 10);
-            // Replace the old password with the new password
-            await dbCollection.updateOne(
-                { email },
-                { $set: { password: hashedPassword } }
-            );
-            // Delete the record associated with the token from the database
-            await passwordResetCollection.deleteOne({ token });
+            if (tokenInDatabase === token) {
+                // Retrieve the user's email address if token matches
+                const email = passwordResetUserRecord.email;
+                // Hash the password entered into reset password
+                const hashedPassword = await bcrypt.hash(receivedData.password, 10);
+                // Replace the old password with the new password
+                await dbCollection.updateOne(
+                    { email },
+                    { $set: { password: hashedPassword } }
+                );
+                // Delete the record associated with the token from the database
+                await passwordResetCollection.deleteOne({ token });
+            } else {
+                response.status(401).json({ error: 'Invalid reset password code.' });
+                return;
+            }
+
+            
+
 
 
             // Send success response message
